Fix building for debian 13 armv7l
In Debian 13 armv7l, time_t and suseconds_t are "long long int" instead
of "long unsigned int" as seen on x86_64 and aarch64. Fix it by casting
to "long unsigned int".

Related: OS#6828

+++ src/ipaccess/ipaccess-config.c
@@ -879,7 +879,7 @@ static void analyze_firmware(const char *filename)
 			printf("\taddr1: 0x%x\n", ntohl(sub_entry->header_entry.addr1));
 			printf("\taddr2: 0x%x\n", ntohl(sub_entry->header_entry.addr2));
 			printf("\tstart: 0x%x\n", ntohl(sub_entry->header_entry.start));
-			printf("\tabs. offset: 0x%lx\n", sub_entry->absolute_offset);
+			printf("\tabs. offset: 0x%lx\n", (unsigned int long) sub_entry->absolute_offset);
 			printf("\n\n");
 
 			dump_entry(sub_entry, part++, fd);
